+++ src/itfPanel.c
@@ -3602,7 +3602,7 @@ u64 FUN_003ace60(u64 param_1,int param_2)
 // FUN_003ACF80 NONMATCHING
 
 
-u64 FUN_003acf80(u64 param_1,int param_2)
+u32 FUN_003acf80(u32 param_1,int param_2)
